requirements.txt
fastapi[standard]
uvicorn[standard]
asyncpg
pydantic
pydantic-settings
pathlib
shortuuid
pyjwt
passlib[bcrypt]
bcrypt
